Reactify for improved sanity
In order to separate the display from the state, everything moved into
either Application or Models. Models holds the Bank and it's children which
contain the current state of the game, plus some spillover in the index.jsx.
Application is all react components and is a direct mapping from the old
uiElements.